Removed persistence
