--- src/persistence.rs
@@ -1,28 +1,0 @@
-use std::{
-    collections::hash_map::DefaultHasher,
-    hash::{Hash, Hasher},
-};
-
-use serde::{Deserialize, Serialize};
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExtractedAttributes {
-    pub id: String,
-    pub content_id: String,
-    pub attributes: serde_json::Value,
-    pub extractor_name: String,
-}
-
-impl ExtractedAttributes {
-    pub fn new(content_id: &str, attributes: serde_json::Value, extractor_name: &str) -> Self {
-        let mut s = DefaultHasher::new();
-        content_id.hash(&mut s);
-        extractor_name.hash(&mut s);
-        let id = format!("{:x}", s.finish());
-        Self {
-            id,
-            content_id: content_id.into(),
-            attributes,
-            extractor_name: extractor_name.into(),
-        }
-    }
-}
